Move yamlcodec to a neutral leaf, breaking the model->persistence inversion
The model layer imported persistence/yamlcodec while persistence imports
model/doc — a mutual layer dependency that inverts the one-way rule
(persistence depends on the domain, never the reverse). yamlcodec is a thin
wrapper over gopkg.in/yaml.v3 with no domain or persistence dependency of its
own (ADR-0020), so it belongs at the leaf like math/, not under persistence.

Relocate persistence/yamlcodec -> top-level yamlcodec/ (import path
oblikovati.org/yamlcodec) and repoint all importers. The domain now reaches
"down" to a neutral leaf instead of "up" the stack.

Guard the inversion so it cannot return: a dedicated transitive test
(TestModelDoesNotImportPersistence, go list -deps ./model/...) plus the
model row in the edge allowlist loses its persistence TODO and gains yamlcodec.
Red-verified by re-adding a persistence import to a model file.

Closes #1615
