+++ head/internal/windowstate/windowstate.go
@@ -11,8 +11,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"oblikovati.org/persistence/yamlcodec"
 	"oblikovati.org/userconfig"
+	"oblikovati.org/yamlcodec"
 )
 
 // State is the saved window placement.
