+++ persistence/viewstate/viewstate.go
@@ -15,8 +15,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"oblikovati.org/persistence/yamlcodec"
 	"oblikovati.org/userconfig"
+	"oblikovati.org/yamlcodec"
 )
 
 // ViewFrame is one view's persisted camera: a name and a look-at frame (eye, target, up in
