+++ yamlcodec/resources.go
